📦 Separate unit tests into steps
