+++ src/commonTest/kotlin/prepared/GameMechanicsTest.kt
@@ -1,3 +1,5 @@
+package prepared
+
 import j4k.candycrush.*
 import j4k.candycrush.GameMechanics.InsertMove
 import j4k.candycrush.GameMechanics.Move
@@ -9,53 +11,6 @@ import kotlin.test.*
  * Test for the [GameMechanics] which performs actions on the [GameField].
  */
 class GameMechanicsTest {
-
-    /**
-     * Test swapping of two tiles ([GameMechanics.swapTiles]).
-     */
-    @Test
-    fun swapTiles() {
-        val field = GameField.fromString(
-            """
-                |[H, H, H, H]
-                |[H, A, B, H]
-                |[H, H, H, H]
-                """.trimMargin()
-        )
-
-        val mechanics = GameMechanics(field)
-        mechanics.swapTiles(Position(1, 1), Position(2, 1))
-
-        assertEquals(
-            """
-                |[H, H, H, H]
-                |[H, B, A, H]
-                |[H, H, H, H]
-                """.trimMargin(), field.toString()
-        )
-    }
-
-    /**
-     * Test remove tile at given position ([GameMechanics.removeTile]).
-     */
-    @Test
-    fun removeTiles() {
-        val field = GameField.fromString(
-            """
-                |[A, A]
-                |[A, A]
-                """.trimMargin()
-        )
-        val mechanics = GameMechanics(field)
-        mechanics.removeTile(Position(0, 0))
-        mechanics.removeTile(Position(1, 1))
-        assertEquals(
-            """
-                |[H, A]
-                |[A, H]
-                """.trimMargin(), field.toString()
-        )
-    }
 
     /**
      * Test if a position points to a tile which is in row with three ([GameMechanics.isInRowWithThree]).
@@ -139,142 +94,6 @@ class GameMechanicsTest {
         assertEquals(emptyList(), mechanics.getConnectedTiles(Position(0, 2)))
     }
 
-    /**
-     * Test [GameMechanics.getVerticalSurroundings] with three tiles in a vertical line.
-     */
-    @Test
-    fun getVerticalSurroundingsWithThreeTilesVertical() {
-        val field = GameField.fromString(
-            """
-                |[H, A, H, H]
-                |[H, A, A, A]
-                |[H, A, H, H]
-                """.trimMargin()
-        )
-        val mechanics = GameMechanics(field)
-
-        // Vertical
-        val vertical: List<TileCell> = listOf(
-            field.getTileCell(1, 0),
-            field.getTileCell(1, 1),
-            field.getTileCell(1, 2)
-        )
-        assertEquals(vertical, mechanics.getVerticalSurroundings(Position(1, 0)))
-        assertEquals(vertical, mechanics.getVerticalSurroundings(Position(1, 1)))
-        assertEquals(vertical, mechanics.getVerticalSurroundings(Position(1, 2)))
-    }
-
-    /**
-     * Test [GameMechanics.getVerticalSurroundings] with three tiles in a horizontal line.
-     */
-    @Test
-    fun getVerticalSurroundingsWithThreeTilesHorizontal() {
-        val field = GameField.fromString(
-            """
-                |[H, A, H, H]
-                |[H, A, A, A]
-                |[H, A, H, H]
-                """.trimMargin()
-        )
-        val mechanics = GameMechanics(field)
-
-        val horizontal: List<TileCell> = listOf(
-            field.getTileCell(1, 1), field.getTileCell(2, 1), field.getTileCell(3, 1)
-        )
-
-        assertEquals(horizontal, mechanics.getHorizontalSurroundings(Position(2, 1)))
-        assertEquals(horizontal, mechanics.getHorizontalSurroundings(Position(1, 1)))
-        assertEquals(horizontal, mechanics.getHorizontalSurroundings(Position(3, 1)))
-    }
-
-    /**
-     * Test [GameMechanics.getHorizontalSurroundings] don't match with holes ([Tile.Hole]).
-     */
-    @Test
-    fun getHorizontalSurroundingsWithHoles() {
-        val field = GameField.fromString(
-            """
-                |[H, H, H]
-                |[H, H, H]
-                |[H, H, H]
-                """.trimMargin()
-        )
-        val mechanics = GameMechanics(field)
-
-        assertEquals(listOf(field.getTileCell(1, 0)), mechanics.getHorizontalSurroundings(Position(1, 0)))
-        assertEquals(listOf(field.getTileCell(1, 1)), mechanics.getHorizontalSurroundings(Position(1, 1)))
-        assertEquals(listOf(field.getTileCell(1, 2)), mechanics.getHorizontalSurroundings(Position(1, 2)))
-    }
-
-    /**
-     * Test [GameMechanics.getHorizontalSurroundings] with six tiles in a horizontal line.
-     */
-    @Test
-    fun getHorizontalSurroundingsWithSixTilesHorizontal() {
-        val field = GameField.fromString(
-            """
-                |[H, A, A, A, A, A, A, H]
-                """.trimMargin()
-        )
-        val mechanics = GameMechanics(field)
-
-        val horizontal: List<TileCell> = listOf(1, 2, 3, 4, 5, 6).map { field.getTileCell(column = it) }
-
-        assertEquals(horizontal, mechanics.getHorizontalSurroundings(Position(column = 1)))
-        assertEquals(horizontal, mechanics.getHorizontalSurroundings(Position(column = 2)))
-        assertEquals(horizontal, mechanics.getHorizontalSurroundings(Position(column = 3)))
-        assertEquals(horizontal, mechanics.getHorizontalSurroundings(Position(column = 4)))
-        assertEquals(horizontal, mechanics.getHorizontalSurroundings(Position(column = 5)))
-        assertEquals(horizontal, mechanics.getHorizontalSurroundings(Position(column = 6)))
-    }
-
-    /**
-     * Test [GameMechanics.getHorizontalSurroundings] with six tiles in a vertical line.
-     */
-    @Test
-    fun getHorizontalSurroundingsWithSixTilesVertical() {
-        val field = GameField.fromString(
-            """
-                |[H]
-                |[A]
-                |[A]
-                |[A]
-                |[A]
-                |[A]
-                |[A]
-                |[H]
-                """.trimMargin()
-        )
-        val mechanics = GameMechanics(field)
-
-        val vertical: List<TileCell> = listOf(1, 2, 3, 4, 5, 6).map { field.getTileCell(row = it) }
-
-        assertEquals(vertical, mechanics.getVerticalSurroundings(Position(row = 1)))
-        assertEquals(vertical, mechanics.getVerticalSurroundings(Position(row = 2)))
-        assertEquals(vertical, mechanics.getVerticalSurroundings(Position(row = 3)))
-        assertEquals(vertical, mechanics.getVerticalSurroundings(Position(row = 4)))
-        assertEquals(vertical, mechanics.getVerticalSurroundings(Position(row = 5)))
-        assertEquals(vertical, mechanics.getVerticalSurroundings(Position(row = 6)))
-    }
-
-    /**
-     * Test [GameMechanics.getVerticalSurroundings] don't match with holes ([Tile.Hole]).
-     */
-    @Test
-    fun getVerticalSurroundingsWithHoles() {
-        val field = GameField.fromString(
-            """
-                |[H, H, H]
-                |[H, H, H]
-                |[H, H, H]
-                """.trimMargin()
-        )
-        val mechanics = GameMechanics(field)
-
-        assertEquals(listOf(field.getTileCell(1, 0)), mechanics.getVerticalSurroundings(Position(1, 0)))
-        assertEquals(listOf(field.getTileCell(1, 1)), mechanics.getVerticalSurroundings(Position(1, 1)))
-        assertEquals(listOf(field.getTileCell(1, 2)), mechanics.getVerticalSurroundings(Position(1, 2)))
-    }
 
     /**
      * Test that a hole does not count as connected with other holes for
