+++ src/commonTest/kotlin/prepared/PositionGridTest.kt
@@ -1,3 +1,5 @@
+package prepared
+
 import com.soywiz.korma.geom.*
 import j4k.candycrush.math.*
 import j4k.candycrush.math.PositionGrid.Position
